# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "简单又好用的消息服务。快来给你\"项目\"添加消息功能。",
        "title": "消息系统 API",
        "termsOfService": "https://github.com/webb-l",
        "contact": {
            "name": "Webb",
            "url": "https://github.com/webb-l",
            "email": "822028533@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1204",
    "basePath": "/",
    "paths": {
        "/message": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "根据用户凭证查询消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "查询消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤语句（title = 标题,status = 0|1|2,...）",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序列（created_at|updated_at|sender_ids|title|content|category|big_content|introducer_ids|status）",
                        "name": "sortColumn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序类型（asc/desc）",
                        "name": "sortType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "查询第几页数据",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息信息",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/response.Message"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationError"
                        }
                    },
                    "401": {
                        "description": "凭证错误",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "502": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "创建消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "创建消息",
                "parameters": [
                    {
                        "description": "创建的数据",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MessageCreateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/response.Message"
                        }
                    },
                    "202": {
                        "description": "创建失败",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationError"
                        }
                    },
                    "401": {
                        "description": "凭证错误",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "502": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "根据数组的数据删除消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "删除消息",
                "parameters": [
                    {
                        "description": "删除的消息",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/request.MessageDeleteRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后返回的数据",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.MessageDeleteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationError"
                        }
                    },
                    "401": {
                        "description": "凭证错误",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "404": {
                        "description": "找不到兑换码",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "系统异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/message/status": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "根据数组的数据更新消息状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "更新状态",
                "parameters": [
                    {
                        "description": "消息状态",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/request.MessageStatusRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后返回的数据",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.MessageStatusResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationError"
                        }
                    },
                    "401": {
                        "description": "凭证错误",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "502": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/message/{id}": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "根据消息id更新消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "更新消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新消息",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MessageCreateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Message"
                        }
                    },
                    "202": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationError"
                        }
                    },
                    "401": {
                        "description": "凭证错误",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "404": {
                        "description": "找不到数据",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "502": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.MessageCreateUpdateRequest": {
            "type": "object",
            "required": [
                "bigContent",
                "category",
                "content",
                "introducerIds",
                "title"
            ],
            "properties": {
                "bigContent": {
                    "type": "string",
                    "example": "复杂的内容"
                },
                "category": {
                    "type": "string",
                    "example": "important"
                },
                "content": {
                    "type": "string",
                    "example": "简单的内容"
                },
                "introducerIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "发给谁"
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "标题"
                }
            }
        },
        "request.MessageDeleteRequest": {
            "type": "object",
            "required": [
                "messageId"
            ],
            "properties": {
                "delete": {
                    "type": "boolean",
                    "example": false
                },
                "messageId": {
                    "type": "string",
                    "example": "id"
                }
            }
        },
        "request.MessageStatusRequest": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "status": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0,
                    "example": 1
                }
            }
        },
        "request.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "字段名",
                    "type": "string"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string"
                },
                "param": {
                    "description": "参数",
                    "type": "string"
                },
                "type": {
                    "description": "字段类型",
                    "type": "string"
                },
                "value": {
                    "description": "字段数值"
                }
            }
        },
        "response.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "response.Message": {
            "type": "object",
            "properties": {
                "big_content": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "introducer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message_id": {
                    "type": "string"
                },
                "sender_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.MessageDeleteResponse": {
            "type": "object",
            "properties": {
                "delete": {
                    "description": "Delete 表示消息是否永久删除。",
                    "type": "boolean"
                },
                "id": {
                    "description": "Id 表示消息的唯一标识符。",
                    "type": "string"
                },
                "status": {
                    "description": "Status 表示消息删除操作的状态，用于指示操作是否成功",
                    "type": "boolean"
                }
            }
        },
        "response.MessageStatusResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Id 表示消息的唯一标识符。",
                    "type": "string"
                },
                "result": {
                    "description": "Result 表示消息状态更新操作的结果。\ntrue 表示更新成功，false 表示更新失败。",
                    "type": "boolean"
                },
                "status": {
                    "description": "Status 表示消息的当前状态。",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}